subeject crud

+++ src/main/java/cl/praxis/bootcapp/services/IBaseServiceCRUD.java
@@ -1,11 +1,13 @@
 package cl.praxis.bootcapp.services;
 
+import cl.praxis.bootcapp.entities.Subject;
+
 import java.util.List;
 
 public interface IBaseServiceCRUD<T> {
     List<T> getAll();
-    T getById(int id);
+    T getById(Long id);
     T create(T t);
     T update(T t);
-    boolean delete(int id);
+    boolean delete(Long id);
 }
